// src/thunder-sdk/src/types/transfer_selector/TransferSelectorAbi.d.ts
export type IdentityInput = Enum<{ Address: AddressInput; ContractId: ContractIdInput }>;
export type IdentityOutput = Enum<{ Address: AddressOutput; ContractId: ContractIdOutput }>;

export type AddressInput = { value: string };
export type AddressOutput = AddressInput;
export type ContractIdInput = { value: string };
export type ContractIdOutput = ContractIdInput;
// ...
interface TransferSelectorAbiInterface extends Interface {
  functions: {
    add_collection_transfer_manager: FunctionFragment;
    get_transfer_manager_1155: FunctionFragment;
    get_transfer_manager_721: FunctionFragment;
    get_transfer_manager_for_token: FunctionFragment;
    initialize: FunctionFragment;
    owner: FunctionFragment;
    remove_collection_transfer_manager: FunctionFragment;
    renounce_ownership: FunctionFragment;
    set_transfer_manager_1155: FunctionFragment;
    set_transfer_manager_721: FunctionFragment;
    transfer_ownership: FunctionFragment;
  };

  encodeFunctionData(functionFragment: "add_collection_transfer_manager", values: [ContractIdInput, ContractIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: "get_transfer_manager_1155", values: []): Uint8Array;
  encodeFunctionData(functionFragment: "get_transfer_manager_721", values: []): Uint8Array;
  encodeFunctionData(functionFragment: "get_transfer_manager_for_token", values: [ContractIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: "initialize", values: [ContractIdInput, ContractIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: "owner", values: []): Uint8Array;
  encodeFunctionData(functionFragment: "remove_collection_transfer_manager", values: [ContractIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: "renounce_ownership", values: []): Uint8Array;
  encodeFunctionData(functionFragment: "set_transfer_manager_1155", values: [ContractIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: "set_transfer_manager_721", values: [ContractIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: "transfer_ownership", values: [IdentityInput]): Uint8Array;

  decodeFunctionData(functionFragment: "add_collection_transfer_manager", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "get_transfer_manager_1155", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "get_transfer_manager_721", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "get_transfer_manager_for_token", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "initialize", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "owner", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "remove_collection_transfer_manager", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "renounce_ownership", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "set_transfer_manager_1155", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "set_transfer_manager_721", data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: "transfer_ownership", data: BytesLike): DecodedValue;
}

export class TransferSelectorAbi extends Contract {
  interface: TransferSelectorAbiInterface;
  functions: {
    add_collection_transfer_manager: InvokeFunction<[collection: ContractIdInput, transfer_manager: ContractIdInput], void>;
    get_transfer_manager_1155: InvokeFunction<[], Option<ContractIdOutput>>;
    get_transfer_manager_721: InvokeFunction<[], Option<ContractIdOutput>>;
    get_transfer_manager_for_token: InvokeFunction<[collection: ContractIdInput], Option<ContractIdOutput>>;
    initialize: InvokeFunction<[transfer_manager_721: ContractIdInput, transfer_manager_1155: ContractIdInput], void>;
    owner: InvokeFunction<[], Option<IdentityOutput>>;
    remove_collection_transfer_manager: InvokeFunction<[collection: ContractIdInput], void>;
    renounce_ownership: InvokeFunction<[], void>;
    set_transfer_manager_1155: InvokeFunction<[erc1155_manager: ContractIdInput], void>;
    set_transfer_manager_721: InvokeFunction<[erc721_manager: ContractIdInput], void>;
    transfer_ownership: InvokeFunction<[new_owner: IdentityInput], void>;
  };
}